fix: resolve stripe checkout review comments
